First phase completed
