+++ src/containers/Homepage.js
@@ -1,11 +1,11 @@
 import React,{Component} from 'react'
-import Home from "../Components/Homepage"
+import Home from "../components/Homepage"
 import PropTypes from 'prop-types';
 import {connect} from 'react-redux'
 import { createMeasure } from '../store/NewMeasure/action'
 
 
-export class Homepages extends Component{
+ class Homepages extends Component{
 
     constructor() {
         super();
@@ -20,14 +20,10 @@ handleChange=(value)=>{
   console.log(value)
 }
 
-submit=(e)=>{
-  
+submit=(e)=>{ 
               e.preventDefault()
               this.props.dispatch(createMeasure(this.state.name))
-               this.props.history.push("/Enterdetails")
-             // window.location="/Enterdetails"
-            
-
+               this.props.history.push("/Enterdetails")                      
 }
 
     render(){
