+++ src/store/rootReducer.js
@@ -1,9 +1,9 @@
 import { combineReducers } from 'redux'
-import todos from './NewMeasure/reducer'
+import measures from './NewMeasure/reducer'
 //import visibilityFilter from './visibilityFilter'
 
 const todoApp = combineReducers({
-  todos
+  measures
 
 })
